refactor: UI관련된 페키지명을 `view`로 변경
