+++ README.md
@@ -62,7 +62,7 @@
 
 ## 리펙터링 요구사항
 
-- [ ] UI 관련한 객체를 `view` 패키지에 구현한다.
+- [X] UI 관련한 객체를 `view` 패키지에 구현한다.
 - [ ] MVC 패턴으로 리팩터링 후의 main 메소드 예시는 다음과 같을 수 있도록 한다.
 
 ```java
