+++ src/main/java/racingcar/controller/RacingCarGame.java
@@ -3,8 +3,8 @@ package racingcar.controller;
 import java.util.function.Supplier;
 import racingcar.domain.Game;
 import racingcar.domain.Vehicles;
-import racingcar.ui.InputView;
-import racingcar.ui.OutputView;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class RacingCarGame {
     public void start() {
