+++ src/main/java/racingcar/view/OutputCommand.java
@@ -1,4 +1,4 @@
-package racingcar.ui;
+package racingcar.view;
 
 public class OutputCommand {
     public static String ASK_CAR_NAMES = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
