+++ src/test/java/racingcar/view/OutputCommandTest.java
@@ -1,4 +1,4 @@
-package racingcar.ui;
+package racingcar.view;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
